42_14/10

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: acesar-m <acesar-m@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/07 09:49:18 by acesar-m          #+#    #+#             */
-/*   Updated: 2024/10/11 11:12:29 by acesar-m         ###   ########.fr       */
+/*   Updated: 2024/10/14 10:06:10 by acesar-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *c);
+size_t	ft_strlen(const char *c);
 
-int	ft_strlen(const char *c)
+size_t	ft_strlen(const char *c)
 {
 	size_t	i;
